parser/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pdfplumber>=0.10.0
pdfminer.six>=20221105
docx2txt~=0.8
odfpy~=1.4
